Stop narrowing the two weight totals whose callees take a long
PetInventory.validateWeight is handed a long count, multiplies it by the item
weight -- a long product -- and accumulates it into an int local. A compound
assignment narrows silently, with no warning from the compiler, and the result
then goes to an overload that takes a long. ItemContainer.validateWeightByItemId,
its counterpart, passes the long product straight through.

SendWareHouseWithDrawList takes both its totals in int and computes the weight
product in int, then hands them to validateWeight and validateCapacity, which
both take a long.

Dismissed in the same sweep, each with what dismisses it: TradeList's private
store total cannot land anywhere but negative when it wraps, because both the
running total and the item product are bounded by MAX_ADENA and twice that is
still short of a full turn, and the negative is tested; RequestBuySeed has the
same test thirty lines below where its overflow check is, and the worst seed the
shipped data allows is a product of 3,936,000; Party squares member levels, at
most 65,025 for a full party; and RequestProcureCropList was already measured
against a crop limit of 9,000 and a reward weight of 2.

+++ java/org/l2jmobius/gameserver/network/clientpackets/SendWareHouseWithDrawList.java
@@ -140,8 +140,10 @@ public class SendWareHouseWithDrawList extends ClientPacket
 			}
 		}
 		
-		int weight = 0;
-		int slots = 0;
+		// validateWeight and validateCapacity both take a long; the totals feeding them were
+		// taken in int, with the product in int as well, so the range was thrown away here.
+		long weight = 0;
+		long slots = 0;
 		for (ItemHolder i : _items)
 		{
 			// Calculate needed slots
@@ -152,7 +154,7 @@ public class SendWareHouseWithDrawList extends ClientPacket
 				return;
 			}
 			
-			weight += i.getCount() * item.getTemplate().getWeight();
+			weight += (long) i.getCount() * item.getTemplate().getWeight();
 			if (!item.isStackable())
 			{
 				slots += i.getCount();
